requirements.txt
pandas
numpy
openai
matplotlib
plotly
langchain
python-dotenv
fastapi
uvicorn
